+++ Adjust/ADJActivityState.h
@@ -18,6 +18,9 @@
 @property (nonatomic, copy) NSString *deviceToken;
 @property (nonatomic, assign) BOOL updatePackages;
 
+@property (nonatomic, copy) NSString *adid;
+@property (nonatomic, strong) NSDictionary *attributionDetails;
+
 // Global counters
 @property (nonatomic, assign) int eventCount;
 @property (nonatomic, assign) int sessionCount;