+++ Adjust/ADJAttribution.m
@@ -12,11 +12,13 @@
 
 @implementation ADJAttribution
 
-+ (ADJAttribution *)dataWithJsonDict:(NSDictionary *)jsonDict {
-    return [[ADJAttribution alloc] initWithJsonDict:jsonDict];
++ (ADJAttribution *)dataWithJsonDict:(NSDictionary *)jsonDict
+                                adid:(NSString *)adid {
+    return [[ADJAttribution alloc] initWithJsonDict:jsonDict adid:adid];
 }
 
-- (id)initWithJsonDict:(NSDictionary *)jsonDict {
+- (id)initWithJsonDict:(NSDictionary *)jsonDict
+                  adid:(NSString *)adid {
     self = [super init];
     if (self == nil) return nil;
 
@@ -31,6 +33,7 @@
     self.adgroup      = [jsonDict objectForKey:@"adgroup"];
     self.creative     = [jsonDict objectForKey:@"creative"];
     self.clickLabel   = [jsonDict objectForKey:@"click_label"];
+    self.adid         = adid;
 
     return self;
 }
@@ -58,6 +61,9 @@
         return NO;
     }
     if (![NSString adjIsEqual:self.clickLabel toString:attribution.clickLabel]) {
+        return NO;
+    }
+    if (![NSString adjIsEqual:self.adid toString:attribution.adid]) {
         return NO;
     }
 
@@ -95,13 +101,17 @@
         [responseDataDic setObject:self.clickLabel forKey:@"click_label"];
     }
 
+    if (self.adid != nil) {
+        [responseDataDic setObject:self.adid forKey:@"adid"];
+    }
+
     return responseDataDic;
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"tt:%@ tn:%@ net:%@ cam:%@ adg:%@ cre:%@ cl:%@",
+    return [NSString stringWithFormat:@"tt:%@ tn:%@ net:%@ cam:%@ adg:%@ cre:%@ cl:%@ adid:%@",
             self.trackerToken, self.trackerName, self.network, self.campaign,
-            self.adgroup, self.creative, self.clickLabel];
+            self.adgroup, self.creative, self.clickLabel, self.adid];
 }
 
 
@@ -136,6 +146,7 @@
         copy.adgroup      = [self.adgroup copyWithZone:zone];
         copy.creative     = [self.creative copyWithZone:zone];
         copy.clickLabel   = [self.clickLabel copyWithZone:zone];
+        copy.adid         = [self.adid copyWithZone:zone];
     }
 
     return copy;
@@ -155,6 +166,7 @@
     self.adgroup      = [decoder decodeObjectForKey:@"adgroup"];
     self.creative     = [decoder decodeObjectForKey:@"creative"];
     self.clickLabel   = [decoder decodeObjectForKey:@"click_label"];
+    self.adid         = [decoder decodeObjectForKey:@"adid"];
 
     return self;
 }
@@ -167,6 +179,7 @@
     [encoder encodeObject:self.adgroup      forKey:@"adgroup"];
     [encoder encodeObject:self.creative     forKey:@"creative"];
     [encoder encodeObject:self.clickLabel   forKey:@"click_label"];
+    [encoder encodeObject:self.adid         forKey:@"adid"];
 }
 
 @end