+++ Adjust/ADJPackageBuilder.h
@@ -21,9 +21,8 @@
 @property (nonatomic, copy) NSDate *purchaseTime;
 
 @property (nonatomic, copy) NSString *deeplink;
-@property (nonatomic, copy) NSString *deviceToken;
 
-@property (nonatomic, strong) NSDictionary *iadDetails;
+@property (nonatomic, strong) NSDictionary *attributionDetails;
 @property (nonatomic, strong) NSDictionary* deeplinkParameters;
 
 @property (nonatomic, copy) ADJAttribution *attribution;