+++ Adjust/Adjust.m
@@ -112,6 +112,14 @@ NSString * const ADJEnvironmentProduction   = @"production";
     [[Adjust getInstance] resetSessionPartnerParameters];
 }
 
++ (ADJAttribution *)attribution {
+    return [[Adjust getInstance] attribution];
+}
+
++ (NSString *)adid {
+    return [[Adjust getInstance] adid];
+}
+
 + (id)getInstance {
     static Adjust *defaultInstance = nil;
     static dispatch_once_t onceToken;
@@ -288,6 +296,16 @@ NSString * const ADJEnvironmentProduction   = @"production";
     [self.sessionParametersActionsArray addObject:^(ADJActivityHandler * activityHandler){
         [activityHandler resetSessionPartnerParametersI:activityHandler];
     }];
+}
+
+- (ADJAttribution *)attribution {
+    if (![self checkActivityHandler]) return nil;
+    return [self.activityHandler attribution];
+}
+
+- (NSString *)adid {
+    if (![self checkActivityHandler]) return nil;
+    return [self.activityHandler adid];
 }
 
 - (void)teardown:(BOOL)deleteState {