+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+### Version 4.11.0 (27th December 2016)
+#### Added
+- Added `adid` field to the attribution callback response.
+- Added accessor `[Adjust adid]` to be able to get `adid` value at any time after obtaining it, not only when session/event callbacks have been triggered.
+- Added accessor `[Adjust attribution]` to be able to get current attribution value at any time after obtaining it, not only when attribution callback has been triggered.
+- Added `AdjustSdkTv` scheme to shared ones in order to allow `Carthage` build for `tvOS`.
+
+#### Changed
+- Updated Criteo plugin:
+    - Added new partner parameter `user_segment` to be sent in `injectUserSegmentIntoCriteoEvents` (for all Criteo events).
+    - Moved `customer_id` to be sent in `injectCustomerIdIntoCriteoEvents` (for all Criteo events).
+    - Added new partner parameter `new_customer` to be sent in `injectTransactionConfirmedIntoEvent`.
+- Firing attribution request as soon as install has been tracked, regardless of presence of attribution callback implementation in user's app.
+- Saveing iAd/AdSearch details to prevent sending duplicated `sdk_click` packages.
+- Updated docs.
+
+#### Fixed
+- Now reading push token value from activity state file when sending package.
+- Fixed memory leak by closing network session.
+- Fixed `TARGET_OS_TV` pre processer check.
+
+---
+
 ### Version 4.10.3 (18th November 2016)
 #### Added
 - Added sending of `os_build` parameter.
@@ -8,7 +31,7 @@
 - It is no longer necessary to have attribution delegate implemented to get deferred deep links.
 - Sending `os_build` or permenent version, not both.
 
---
+---
 
 ### Version 4.10.2 (30th September 2016)
 #### Fixed